Move more stuff to renderer_utils
